Turn into PHPCS based sniff
