+++ includes/Checker/Default_Check_Repository.php
@@ -59,7 +59,7 @@ class Default_Check_Repository extends Empty_Check_Repository {
 				'no_unfiltered_uploads'      => new Checks\Plugin_Repo\No_Unfiltered_Uploads_Check(),
 				'trademarks'                 => new Checks\Plugin_Repo\Trademarks_Check(),
 				'non_blocking_scripts'       => new Checks\Performance\Non_Blocking_Scripts_Check(),
-				'remote_files'               => new Checks\Plugin_Repo\Plugin_Remote_Files(),
+				'offloading_files'           => new Checks\Plugin_Repo\Offloading_Files(),
 			)
 		);
