+++ phpcs-sniffs/PluginCheck/ruleset.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<ruleset xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" name="Plugin Check" namespace="PluginCheck\Sniffs" xsi:noNamespaceSchemaLocation="https://raw.githubusercontent.com/squizlabs/PHP_CodeSniffer/master/phpcs.xsd">
+
+	<description>Plugin Check Sniffs</description>
+
+	<rule ref="PluginCheck.EnqueuedResourceOffloadingSniff"/>
+	<rule ref="PluginCheck.OffloadingSniff"/>
+
+</ruleset>
