+++ phpcs-sniffs/composer.json
@@ -1,0 +1,72 @@
+{
+  "name": "plugin-check/phpcs-sniffs",
+  "type": "phpcodesniffer-standard",
+  "description": "PHP_CodeSniffer rules (sniffs) for static checks in Plugin Check",
+  "keywords": [
+    "phpcs",
+    "standards",
+    "static analysis",
+    "plugin-check"
+  ],
+  "license": "GPL-2.0-or-later",
+  "authors": [
+    {
+      "name": "Contributors",
+      "homepage": "https://github.com/WordPress/plugin-check/graphs/contributors"
+    }
+  ],
+  "require": {
+    "php": ">=7.0",
+    "ext-libxml": "*",
+    "ext-tokenizer": "*",
+    "ext-xmlreader": "*",
+    "squizlabs/php_codesniffer": "^3.7.2",
+    "phpcsstandards/phpcsutils": "^1.0.8"
+  },
+  "require-dev": {
+    "phpcompatibility/php-compatibility": "^9.0",
+    "phpunit/phpunit": "^5.0 || ^6.0 || ^7.0",
+    "php-parallel-lint/php-parallel-lint": "^1.3.2",
+    "php-parallel-lint/php-console-highlighter": "^1.0.0",
+    "wp-coding-standards/wpcs": "^3.0"
+  },
+  "suggest": {
+    "ext-mbstring": "For improved results"
+  },
+  "config": {
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    }
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-main": "2.12.x-dev"
+    }
+  },
+  "scripts": {
+    "lint": [
+      "@php ./vendor/php-parallel-lint/php-parallel-lint/parallel-lint . -e php --show-deprecated --exclude vendor --exclude .git"
+    ],
+    "check-cs": [
+      "@php ./vendor/squizlabs/php_codesniffer/bin/phpcs"
+    ],
+    "fix-cs": [
+      "@php ./vendor/squizlabs/php_codesniffer/bin/phpcbf"
+    ],
+    "run-tests": [
+      "@php ./vendor/phpunit/phpunit/phpunit --filter PluginCheck ./vendor/squizlabs/php_codesniffer/tests/AllTests.php --no-coverage"
+    ],
+    "check-all": [
+      "@lint",
+      "@check-cs",
+      "@run-tests"
+    ]
+  },
+  "scripts-descriptions": {
+    "lint": "Lint PHP files against parse errors.",
+    "check-cs": "Run the PHPCS script against the entire codebase.",
+    "fix-cs": "Run the PHPCBF script to fix all the autofixable violations on the codebase.",
+    "run-tests": "Run all the unit tests for the Plugin Check Coding Standards sniffs.",
+    "check-all": "Run all checks (lint, phpcs) and tests."
+  }
+}
